refactor : DEFAULT_IMAGE 리팩토링
