+++ client/src/constants/defaultImage.ts
@@ -1,0 +1,3 @@
+import defaultImage from "/assets/images/logo-profile2.png";
+
+export const DEFAULT_IMAGE = defaultImage;
