+++ client/src/constants/domains.ts
@@ -1,0 +1,10 @@
+export const allowedDomains = [
+  "naver.com",
+  "github.com",
+  "yahoo.com",
+  "daum.net",
+  "kakao.com",
+  "routrip.com",
+  "gmail.com",
+  "test.com",
+];
