+++ client/src/pages/LoginPage.tsx
@@ -1,12 +1,13 @@
-import { JoinPageStyle, WrapperStyle, allowedDomains, placeholderHander } from "./JoinPage";
+import { JoinPageStyle, WrapperStyle, placeholderHander } from "./JoinPage";
 import Title from "@/components/common/Title";
 import { LoginProps, useAuth } from "@/hooks/useAuth";
 import icons from "@/icons/icons";
 import { useForm } from "react-hook-form";
 import InputText from "@/components/common/Input";
-import { emailOptions, passwordOptions } from "@/config/registerOptions";
+import { emailOptions } from "@/config/registerOptions";
 import { Button } from "@/components/common/Button";
 import { Link } from "react-router-dom";
+import { allowedDomains } from "@/constants/domains";
 
 const LoginPage = () => {
   const { userLogin } = useAuth();
