+++ client/src/pages/PostDetailPage.tsx
@@ -3,8 +3,8 @@ import { addNewSchedule } from "@/apis/schedule.api";
 import { Button } from "@/components/common/Button";
 import PlaceModal from "@/components/common/PlaceModal";
 import PostCommentCard from "@/components/common/PostComment";
-import { DEFAULT_IMAGE } from "@/components/common/ProfileCard";
 import WriteTopBtn from "@/components/common/WriteTopBtn";
+import { DEFAULT_IMAGE } from "@/constants/defaultImage";
 import icons from "@/icons/icons";
 import { PostComment } from "@/models/comment.model";
 import { PlaceDetails } from "@/models/place.model";
